Strengthen the empty query checks
